% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the PGA-cmdr application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny]{shinyApp}}.}
}
\value{
A \code{shiny.appobj} (invisibly).
}
\description{
Run the PGA-cmdr application
}
\examples{
\dontrun{
  run_app()
}
}
