% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinypopgen-package.R
\docType{package}
\name{shinypopgen-package}
\alias{shinypopgen}
\alias{shinypopgen-package}
\title{shinypopgen: Interactive Shiny Application for Population Genetics}
\description{
ShinyPopGen is an interactive Shiny application for exploratory and descriptive population genetics analyses from individual-based multilocus datasets. It supports data import and formatting, marker inspection, estimation of standard diversity and differentiation statistics (Weir & Cockerham 1984), linkage disequilibrium analyses, and null allele screening.
}
\author{
\strong{Maintainer}: Naffiou Kadiri \email{naffiou.kadiri@ird.fr}

Other contributors:
\itemize{
  \item INTERTRYP / IRD / CIRAD [copyright holder]
}

}
\keyword{internal}
